Rename notification logic module to avoid case-only filename collision
notificationCenter.ts and NotificationCenter.tsx were distinct only by case
on macOS's case-insensitive filesystem, causing import resolution issues.
Renamed notificationCenter.ts → notificationLogic.ts, updated imports in the
component and test file, and fixed App.tsx to use an extensionless import.
